refactoring codebase: move routes from server file to index.js
