+++ routes/index.js
@@ -1,5 +1,16 @@
-const router = require('express').Router()
+const bcrypt = require('bcrypt-nodejs')
 const database = require('./../database')
+const moment = require('moment')
+const router = require('express').Router()
+const passport = require('./../config/login')
+
+const userCheck = (request, response, next) => {
+  if (request.user) {
+    return next()
+  } else {
+    response.redirect('/login')
+  }
+}
 
 router.get('/', (request, response) => {
   database.getAlbums((error, albums) => {
@@ -20,6 +31,63 @@ router.get('/albums/:albumID', (request, response) => {
     } else {
       const album = albums[0]
       response.render('album', { album: album })
+    }
+  })
+})
+
+router.get('signup', (request, response) => {
+  response.render('signup')
+})
+
+router.post('/signup', (request, response) => {
+  let cryptPword = bcrypt.hashSync(request.body.password)
+  let date_joined = moment().format('MM-DD-YYYY')
+  let certificate = [request.body.name, request.body.email, cryptPword, date_joined]
+
+  database.createUser(certificate, (error) => {
+    if (error) {
+      response.render('signup', { user: null })
+    } else {
+      response.redirect('login')
+    }
+  })
+})
+
+router.get('/login', (request, response) => {
+  let user = request.user ? request.user[0] : null
+  response.render('login', {
+    user: request.user
+  })
+})
+
+router.post('/login',
+  passport.authenticate('login', {
+    successRedirect: '/profile',
+    failureRedirect: '/login'
+  }))
+
+router.get('/logout', (request, response) => {
+  request.session.destroy(() => {
+    response.redirect('/')
+  })
+})
+
+router.get('/profile', userCheck, (request, response) => {
+  let { user } = request.session.passport
+  database.getUserByID(user, (error, id) => {
+    if (error) {
+      response.status(500).render('error', { error: error })
+    } else {
+      let userID = id[0].id
+      database.getUserReviews(userID, (error) => {
+        if (error) {
+          response.status(500).render('error', { error: error })
+        } else {
+          response.render('profile', {
+            user: request.user
+          })
+        }
+      })
     }
   })
 })
