+++ server.js
@@ -2,12 +2,9 @@ const express = require('express')
 const app = express()
 const bodyParser = require('body-parser')
 const cookieParser = require('cookie-parser')
-const database = require('./database')
+const passport = require('./config/login')
 const routes = require('./routes')
-const passport = require('./config/passport')
 const session = require('express-session')
-const bcrypt = require('bcrypt-nodejs')
-const moment = require('moment')
 
 require('ejs')
 app.set('view engine', 'ejs');
@@ -16,54 +13,11 @@ app.use(express.static('public'))
 app.use(cookieParser())
 app.use(bodyParser.urlencoded({ extended: false }))
 
-app.use('/', routes)
-
 app.use(session({ secret: 'notsosecret', resave: false, saveUninitialized: false }))
 app.use(passport.initialize())
 app.use(passport.session())
 
-app.get('/signup', (request, response) => {
-  response.render('signup')
-})
-
-app.post('/signup', (request, response) => {
-  let cryptPword = bcrypt.hashSync(request.body.password)
-  let date_joined = moment().format('MM-DD-YYYY')
-  let certificate = [request.body.name, request.body.email, cryptPword, date_joined]
-
-  database.createUser(certificate, (error) => {
-    if (error) {
-      response.render('signup', { user: null })
-    } else {
-      response.redirect('login')
-    }
-  })
-})
-
-app.get('/login', (request, response) => {
-  let user = request.user ? request.user[0] : null
-  response.render('login', {
-    user: request.user
-  })
-})
-
-app.post('/login',
-  passport.authenticate('login', {
-    successRedirect: '/users',
-    failureRedirect: '/login'
-  }))
-
-app.get('/logout', (request, response) => {
-  request.session.destroy(() => {
-    response.redirect('/')
-  })
-})
-
-app.get('/users', (request, response) => {
-  response.render('users', {
-    user: request.user
-  })
-})
+app.use('/', routes)
 
 app.use((request, response) => {
   response.status(404).render('not_found')
